+++ mods/BrewLAN/hook/lua/ui/help/unitdescription.lua
@@ -177,6 +177,7 @@ Description['srb4402'] = "<LOC Unit_Description_BrewLAN_T4B006>Experimental coun
 Description['ssb5401'] = "<LOC Unit_Description_BrewLAN_T4B007>Experimental quantum gateway device. Capable of instantly transporting units across the map to other gates."
 Description['seb3404'] = "<LOC Unit_Description_BrewLAN_T4B008>Experimental sensor array. Displays the location of enemy units. Upkeep costs per-unit being watched. Sensors can't penetrate areas of heavy sensor-distruption."
 Description['sab0401'] = "<LOC Unit_Description_BrewLAN_T4B009>Constructs air units."
+Description['srb0401'] = "<LOC Unit_Description_BrewLAN_T4B010>Constructs land units."
 
 --Megalith eggs
 Description['srl0001'] = "<LOC Unit_Description_0152>Amphibious destroyer. Armed with a single Dual-Proton cannon, AA auto-cannon and torpedo tubes."